@@ -14,7 +14,7 @@ require 'ontologies_api_client'
 
 class ApplicationController < ActionController::Base
   helper :all # include all helpers, all the time
-  helper_method :bp_config_json # export this method to helpers
+  helper_method :bp_config_json, :using_captcha?
 
   # Pull configuration parameters for REST connection.
   REST_URI = $REST_URL
@@ -86,7 +86,7 @@ class ApplicationController < ActionController::Base
 
   def not_found
     if request.xhr?
-      render text: "Error: load failed"
+      render plain: "Error: load failed"
       return
     end
     
