@@ -28,8 +28,8 @@
 //= require bp_admin
 //= require bp_recommender
 //= require bp_property_tree
-//= require bp_ont_chart
 //= require home
+//= require ontologies
 //= require projects
 //= require submissions
 
